pip_requirements.txt
numpy
pronouncing
pytorch-nlp